Prioritize checking of site-specific rules in detection phase.
This can drastically reduce the number of detection steps in cases where
we have site-specific rules. If we have N rules, and M check iterations
to detect a rule, this reduces the number of checks from N*M to 1+(N*M-1).
In most cases with site-specific rules, where detection will happen in
the first 2 iterations, this is a >=50% reduction in checks.